Room decoration: Use shields instead of padlocks
#2906
