+++ CHANGES.rst
@@ -5,6 +5,7 @@ Improvements:
  * ON/OFF Cross-signing development in a Lab setting (#2855).
  * RoomVC: Update encryption decoration with shields (#2934, #2930, #2906).
  * Settings: Remove "End-to-End Encryption" from the LABS section (#2941).
+ * Room decoration: Use shields instead of padlocks (#2906).
 
 Changes in 0.10.4 (2019-12-11)
 ===============================================
