+++ Riot/AppDelegate.m
@@ -266,6 +266,7 @@ NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey = @"AppDe
     // Set the App Group identifier.
     MXSDKOptions *sdkOptions = [MXSDKOptions sharedInstance];
     sdkOptions.applicationGroupIdentifier = @"group.im.vector";
+    sdkOptions.computeE2ERoomSummaryTrust = YES;
 
     // Redirect NSLogs to files only if we are not debugging
     if (!isatty(STDERR_FILENO))
