+++ Riot/Modules/Common/Recents/Views/RecentTableViewCell.m
@@ -137,7 +137,15 @@ static const CGFloat kDirectRoomBorderWidth = 3.0;
         
         self.directRoomBorderView.hidden = !roomCellData.roomSummary.room.isDirect;
 
-        self.encryptedRoomIcon.hidden = !roomCellData.roomSummary.isEncrypted;
+        if (roomCellData.roomSummary.isEncrypted)
+        {
+            self.encryptedRoomIcon.hidden = NO;
+            self.encryptedRoomIcon.image = [self shieldImageForTrustLevel:roomCellData.roomSummary.roomEncryptionTrustLevel];
+        }
+        else
+        {
+            self.encryptedRoomIcon.hidden = YES;
+        }
 
         [roomCellData.roomSummary setRoomAvatarImageIn:self.roomAvatar];
     }
@@ -151,6 +159,34 @@ static const CGFloat kDirectRoomBorderWidth = 3.0;
 {
     // The height is fixed
     return 74;
+}
+
+- (UIImage*)shieldImageForTrustLevel:(RoomEncryptionTrustLevel)roomEncryptionTrustLevel
+{
+    UIImage *shieldImage;
+    
+    NSString *encryptionIconName;
+    switch (roomEncryptionTrustLevel)
+    {
+        case RoomEncryptionTrustLevelWarning:
+            encryptionIconName = @"encryption_warning";
+            break;
+        case RoomEncryptionTrustLevelNormal:
+            encryptionIconName = @"encryption_normal";
+            break;
+        case RoomEncryptionTrustLevelTrusted:
+            encryptionIconName = @"encryption_trusted";
+            break;
+        case RoomEncryptionTrustLevelUnknown:
+            encryptionIconName = @"encryption_normal";
+            break;
+    }
+    
+    if (encryptionIconName)
+    {
+        shieldImage = [UIImage imageNamed:encryptionIconName];
+    }
+    return shieldImage;
 }
 
 @end
