+++ Riot/Modules/Home/Views/RoomCollectionViewCell.m
@@ -159,7 +159,15 @@ static const CGFloat kDirectRoomBorderWidth = 3.0;
         
         self.directRoomBorderView.hidden = !roomCellData.roomSummary.room.isDirect;
         
-        self.encryptedRoomIcon.hidden = !roomCellData.roomSummary.isEncrypted;
+        if (roomCellData.roomSummary.isEncrypted)
+        {
+            self.encryptedRoomIcon.hidden = NO;
+            self.encryptedRoomIcon.image = [self shieldImageForTrustLevel:roomCellData.roomSummary.roomEncryptionTrustLevel];
+        }
+        else
+        {
+            self.encryptedRoomIcon.hidden = YES;
+        }
         
         [roomCellData.roomSummary setRoomAvatarImageIn:self.roomAvatar];
     }
@@ -205,6 +213,34 @@ static const CGFloat kDirectRoomBorderWidth = 3.0;
         return roomCellData.roomSummary.roomId;
     }
     return nil;
+}
+
+- (UIImage*)shieldImageForTrustLevel:(RoomEncryptionTrustLevel)roomEncryptionTrustLevel
+{
+    UIImage *shieldImage;
+    
+    NSString *encryptionIconName;
+    switch (roomEncryptionTrustLevel)
+    {
+        case RoomEncryptionTrustLevelWarning:
+            encryptionIconName = @"encryption_warning";
+            break;
+        case RoomEncryptionTrustLevelNormal:
+            encryptionIconName = @"encryption_normal";
+            break;
+        case RoomEncryptionTrustLevelTrusted:
+            encryptionIconName = @"encryption_trusted";
+            break;
+        case RoomEncryptionTrustLevelUnknown:
+            encryptionIconName = @"encryption_normal";
+            break;
+    }
+    
+    if (encryptionIconName)
+    {
+        shieldImage = [UIImage imageNamed:encryptionIconName];
+    }
+    return shieldImage;
 }
 
 @end
